test(connectors): align gateway test filter
Move the provider-config deserialization test under connectors::tests so local validation's path-derived nextest filter selects it instead of failing with an empty test set.
